start improving readfile

+++ include/rating_data.hpp
@@ -8,7 +8,7 @@ using namespace std;
 
 struct MovieRating {
     int movieId;
-    double rating;
+    float rating;
 };
 
 string get_formated_user_line(int userId, const std::vector<MovieRating>& ratings);
